+++ rust-src/concordium_base/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Removed `ChainParameterVersionX` types and the `MintDistributionFamily`, `GASRewardsFamily` and `AuthorizationsFamily` traits and their implementations.
 - Revised `UpdateSigner` implementations not to use references, since the one method (`sign_update_hash`) already takes a reference (`&self`).
 - Made `find_authorized_keys` public for easier re-use.
+- Changed value inside `AttributeKind` to be non-public since it has a length constraint. Implemented `AttributeKind::try_new`, `AsRef<str> for AttributeKind` and `From<AttributeKind> for String`
+  to allow constructing and accessing values.
 - Introduced the trait `StructuredDigest` to add data to `RandomOracle` and other hashes
 - Removed the method `RandomOracle::add` and deprecated `RandomOracle::extend_from`