+++ rust-src/concordium_base/CHANGELOG.md
@@ -6,6 +6,15 @@
 - Made `find_authorized_keys` public for easier re-use.
 - Changed value inside `AttributeKind` to be non-public since it has a length constraint. Implemented `AttributeKind::try_new`, `AsRef<str> for AttributeKind` and `From<AttributeKind> for String`
   to allow constructing and accessing values.
+- Added `Eq` and `PartialEq` instances for
+  - `UpdateKeysThreshold`
+  - `InclusiveRange<T>`
+  - `LeverageFactor`
+  - `MintRate`
+  - `CapitalBound`
+  - `Ratio`
+  - `HigherLevelAccessStructure<Kind>`
+  - `AccessStructure`
 
 ## 8.0.0-alpha.3 (2025-10-08)